Complete react-multiple-components - cleaned up buttons into 1 component
