--- react-multiple-components/my-app/src/PrevNextButtons.tsx
@@ -1,8 +1,0 @@
-type Props = {
-  text: string;
-  onClick?: () => void;
-};
-
-export function PrevNextButtons({ text, onClick }: Props) {
-  return <button onClick={onClick}>{text}</button>;
-}
